integrate actual data into map view
for tripMapView, each pin would represent a trip. the data is obtained from the cover photo of the trip.
for the tripDetailView, each pin represents a photo.
more cleanup is needed to make the map cluster information and make sure that pins on the detail view link to an imageview